ex1 quarkus/kafka - fin

+++ 22_quarkus/demo_kafka/micros/payment-service/src/main/java/org/example/client/AccountServiceClient.java
@@ -13,5 +13,5 @@ public interface AccountServiceClient {
 
     @GET
     @Path("/balance-check/{idClient}/{amount}")
-    CheckBalanceDto checkBalance(@PathParam("idClient") long idClient, @PathParam("amount") float amount);
+    CheckBalanceDto checkBalance(@PathParam("idClient") long idClient, @PathParam("amount") double amount);
 }
